requirements.txt
django
djangorestframework
psycopg2
drf-flex-fields
djangorestframework-simplejwt
